+++ solid.h
@@ -25,12 +25,12 @@
 
 struct Solid {
     struct {
-        uint32_t num;
-        uint32_t size;
+        uint64_t num;
+        uint64_t size;
     } faces;
 
     float* vertices;
-    uint32_t* elements;
+    uint64_t* elements;
     float* colors;
     float* normals;
     float* texcoords;
@@ -44,7 +44,7 @@ struct Tetrahedron {
     struct Solid solid;
 
     float vertices[36];
-    uint32_t elements[12];    
+    uint64_t elements[12];    
     float colors[48];
     float normals[36];
     float texcoords[24];
@@ -56,7 +56,7 @@ struct Cube {
     struct Solid solid;
 
     float vertices[108];
-    uint32_t elements[36];    
+    uint64_t elements[36];    
     float colors[144];
     float normals[108];
     float texcoords[72];
@@ -68,7 +68,7 @@ struct Sphere16 {
     struct Solid solid;
 
     float vertices[(16*6*2+16*2)*3*3];
-    uint32_t elements[16*6*2*3+16*3*2];    
+    uint64_t elements[16*6*2*3+16*3*2];    
     float colors[(16*6*2+16*2)*4*3];
     float normals[(16*6*2+16*2)*3*3];
     float texcoords[(16*6*2+16*2)*2*3];
@@ -80,7 +80,7 @@ struct Sphere32 {
     struct Solid solid;
 
     float vertices[(32*14*2+32*2)*3*3];
-    uint32_t elements[32*14*2*3+32*3*2];    
+    uint64_t elements[32*14*2*3+32*3*2];    
     float colors[(32*14*2+32*2)*4*3];
     float normals[(32*14*2+32*2)*3*3];
     float texcoords[(32*14*2+32*2)*2*3];
